Style(Dummy): 반복되는 코드 제거 및 파일 분리
